requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pysnmp==4.4.12
pyasn1==0.4.8
netmiko==4.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
python-dotenv==0.19.0
alembic==1.7.3
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
redis==4.3.4
fastapi-cache2[redis]==0.1.6
aioredis==2.0.1